fix: insert 쿼리 뒤의 update 쿼리 없애기
