+++ src/main/java/com/kusitms/wannafly/command/applicationform/domain/ApplicationItemRepository.java
@@ -1,0 +1,6 @@
+package com.kusitms.wannafly.command.applicationform.domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ApplicationItemRepository extends JpaRepository<ApplicationItem, Long> {
+}
