+++ src/test/java/com/kusitms/wannafly/command/applicationform/application/ApplicationFormServiceTest.java
@@ -102,42 +102,6 @@ class ApplicationFormServiceTest extends ServiceTest {
         }
     }
 
-    @DisplayName("지원서의 지원 항목을 추가할 때")
-    @Nested
-    class AddItemTest {
-
-        @Test
-        void 로그인_회원이_지원서_작성자면_추가_가능하다() {
-            // given
-            Long formId = applicationFormService.createForm(loginMember, FORM_CREATE_REQUEST);
-
-            // when
-            Long itemId = applicationFormService.addItem(formId, loginMember, ITEM_CREATE_REQUEST);
-
-            // then
-            ApplicationFormResponse form = applicationFormQueryService.findOne(formId, loginMember);
-            assertAll(
-                    () -> assertThat(itemId).isEqualTo(4),
-                    () -> assertThat(form.applicationItems()).hasSize(4)
-            );
-        }
-
-        @Test
-        void 로그인_회원이_지원서_작성자가_아니면_예외가_발생한다() {
-            // given
-            Long formId = applicationFormService.createForm(loginMember, FORM_CREATE_REQUEST);
-
-            // when then
-            LoginMember requester = new LoginMember(2L);
-            assertThatThrownBy(() -> applicationFormService.addItem(
-                    formId, requester, ITEM_CREATE_REQUEST)
-            )
-                    .isInstanceOf(BusinessException.class)
-                    .extracting("errorCode")
-                    .isEqualTo(ErrorCode.INVALID_WRITER_OF_FORM);
-        }
-    }
-
     @DisplayName("지원서를 삭제할 때")
     @Nested
     class DeleteTest {
